+++ clients/common/near.cpp
@@ -24,7 +24,7 @@
 
 #include "near.h"
 #include "hipblas.h"
-#include "hipblas_vector.hpp"
+#include "host_vector.hpp"
 #include "utility.h"
 
 /* ========================================Gtest Unit Check
