+++ clients/include/device_strided_batch_matrix.hpp
