+++ clients/include/device_strided_batch_vector.hpp
