+++ clients/include/host_batch_matrix.hpp
@@ -153,7 +153,7 @@ public:
     //!
     //! @brief Constant cast to a double pointer.
     //!
-    operator const T* const *()
+    operator const T* const*()
     {
         return m_data;
     }
