+++ clients/include/host_batch_vector.hpp
@@ -155,7 +155,7 @@ public:
     //!
     //! @brief Constant cast to a double pointer.
     //!
-    operator const T* const *()
+    operator const T* const*()
     {
         return this->m_data;
     }
