+++ clients/include/host_strided_batch_matrix.hpp
