+++ clients/include/host_strided_batch_vector.hpp
