+++ clients/include/near.h
@@ -26,7 +26,7 @@
 #define _NEAR_H
 
 #include "hipblas.h"
-#include "hipblas_vector.hpp"
+#include "host_vector.hpp"
 
 #ifdef GOOGLE_TEST
 #include "gtest/gtest.h"
@@ -101,33 +101,35 @@ template <class Tc, class Ti, class To>
 static constexpr double sum_error_tolerance_for_gfx11 = 0.0;
 
 template <>
-HIPBLAS_CLANG_STATIC constexpr double
-    sum_error_tolerance_for_gfx11<float, hipblasBfloat16, float> = 1 / 10.0;
+HIPBLAS_CLANG_STATIC constexpr double sum_error_tolerance_for_gfx11<float, hipblasBfloat16, float>
+    = 1 / 10.0;
 
 template <>
 HIPBLAS_CLANG_STATIC constexpr double
-    sum_error_tolerance_for_gfx11<float, hipblasBfloat16, hipblasBfloat16> = 1 / 10.0;
+    sum_error_tolerance_for_gfx11<float, hipblasBfloat16, hipblasBfloat16>
+    = 1 / 10.0;
+
+template <>
+HIPBLAS_CLANG_STATIC constexpr double sum_error_tolerance_for_gfx11<float, hipblasHalf, float>
+    = 1 / 100.0;
+
+template <>
+HIPBLAS_CLANG_STATIC constexpr double sum_error_tolerance_for_gfx11<float, hipblasHalf, hipblasHalf>
+    = 1 / 100.0;
 
 template <>
 HIPBLAS_CLANG_STATIC constexpr double
-    sum_error_tolerance_for_gfx11<float, hipblasHalf, float> = 1 / 100.0;
+    sum_error_tolerance_for_gfx11<hipblasHalf, hipblasHalf, hipblasHalf>
+    = 1 / 100.0;
 
 template <>
 HIPBLAS_CLANG_STATIC constexpr double
-    sum_error_tolerance_for_gfx11<float, hipblasHalf, hipblasHalf> = 1 / 100.0;
+    sum_error_tolerance_for_gfx11<hipblasComplex, hipblasComplex, hipblasComplex>
+    = 1 / 10000.0;
 
 template <>
 HIPBLAS_CLANG_STATIC constexpr double
-    sum_error_tolerance_for_gfx11<hipblasHalf, hipblasHalf, hipblasHalf> = 1 / 100.0;
-
-template <>
-HIPBLAS_CLANG_STATIC constexpr double
-    sum_error_tolerance_for_gfx11<hipblasComplex, hipblasComplex, hipblasComplex> = 1 / 10000.0;
-
-template <>
-HIPBLAS_CLANG_STATIC constexpr double
-    sum_error_tolerance_for_gfx11<hipblasDoubleComplex,
-                                  hipblasDoubleComplex,
-                                  hipblasDoubleComplex> = 1 / 1000000.0;
+    sum_error_tolerance_for_gfx11<hipblasDoubleComplex, hipblasDoubleComplex, hipblasDoubleComplex>
+    = 1 / 1000000.0;
 
 #endif
