+++ clients/include/norm.h
@@ -26,7 +26,7 @@
 #define _NORM_H
 
 #include "hipblas.h"
-#include "hipblas_vector.hpp"
+#include "host_vector.hpp"
 #include "utility.h"
 
 /* =====================================================================
