+++ clients/include/testing_common.hpp
@@ -36,20 +36,20 @@
 #else
 #include "hipblas_no_fortran.hpp"
 #endif
-#include "device_batched_matrix.hpp"
-#include "device_batched_vector.hpp"
+#include "device_batch_matrix.hpp"
+#include "device_batch_vector.hpp"
 #include "device_matrix.hpp"
-#include "device_strided_batched_matrix.hpp"
-#include "device_strided_batched_vector.hpp"
+#include "device_strided_batch_matrix.hpp"
+#include "device_strided_batch_vector.hpp"
 #include "device_vector.hpp"
 #include "hipblas_init.hpp"
 #include "hipblas_test.hpp"
-#include "hipblas_vector.hpp"
-#include "host_batched_matrix.hpp"
-#include "host_batched_vector.hpp"
+#include "host_batch_matrix.hpp"
+#include "host_batch_vector.hpp"
 #include "host_matrix.hpp"
-#include "host_strided_batched_matrix.hpp"
-#include "host_strided_batched_vector.hpp"
+#include "host_strided_batch_matrix.hpp"
+#include "host_strided_batch_vector.hpp"
+#include "host_vector.hpp"
 #include "near.h"
 #include "norm.h"
 #include "testing_macros.hpp"
